# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "FitCzar Fitness API",
        "description": "FitCzar is a Web application for assisting users as they pursue customized health and fitness goals. It will track progress, make recommendations, accept user choices and updates, and optionally connect users with resources to support their health and fitness journey."
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter 'Bearer' [space] and then your JWT token in the text input below."
            }
        },
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "height": {
                        "type": "number"
                    },
                    "weight": {
                        "type": "number"
                    },
                    "fitnessGoals": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "medicalConditions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "dietaryRestrictions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "workoutHistory": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "firstName",
                    "lastName",
                    "phoneNumber",
                    "email",
                    "password",
                    "age",
                    "gender",
                    "height",
                    "weight"
                ]
            },
            "UserResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    },
                    "result": {
                        "$ref": "#/components/schemas/User"
                    }
                }
            },
            "Login": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "expiresIn": {
                        "type": "integer"
                    },
                    "userId": {
                        "type": "string"
                    }
                }
            },
            "Workout": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "exercises": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "sets": {
                                    "type": "integer"
                                },
                                "reps": {
                                    "type": "integer"
                                },
                                "weight": {
                                    "type": "number"
                                },
                                "duration": {
                                    "type": "integer"
                                }
                            },
                            "required": [
                                "name"
                            ]
                        }
                    },
                    "creator": {
                        "type": "string"
                    }
                },
                "required": [
                    "title",
                    "description",
                    "exercises",
                    "creator"
                ]
            },
            "WorkoutResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "workout": {
                        "$ref": "#/components/schemas/Workout"
                    }
                }
            }
        }
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "summary": "User registration",
                "description": "Create a new user account",
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, user already registered or validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "summary": "User login",
                "description": "Authenticate a user and return a JWT token",
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Login"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User authenticated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "summary": "User logout",
                "description": "Logs out the user by invalidating the token.",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Authentication failed"
                    },
                    "500": {
                        "description": "Server error"
                      }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "summary": "Get user by ID",
                "description": "Retrieve user information by ID",
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "User ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Fetching user failed"
                    }
                }
            },
            "delete": {
                "summary": "Delete user by ID",
                "description": "Delete a user by their ID",
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "User ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Deleting user failed"
                    }
                }
            }
        },
        "/api/user/update-profile/{id}": {
            "put": {
                "summary": "Update user profile by ID",
                "description": "Update a user's profile by their ID",
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "User ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Updating user failed"
                    }
                }
            }
        },
        "/api/workouts/{id}": {
            "post": {
                "summary": "Create workout",
                "description": "Create a new workout for an authenticated user",
                "tags": [
                    "Workouts"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Workout ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Workout"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Workout created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkoutResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "patch": {
                "summary": "Update workout",
                "description": "Update an existing workout by ID for an authenticated user",
                "tags": [
                    "Workouts"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Workout ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "propName": {
                                            "type": "string"
                                        },
                                        "value": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "propName",
                                        "value"
                                    ]
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Workout updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkoutResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Workout not found or not authorized to update"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "summary": "Delete workout",
                "description": "Delete an existing workout by ID for an authenticated user",
                "tags": [
                    "Workouts"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Workout ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workout deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkoutResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Workout not found or not authorized to delete"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/workouts": {
            "get": {
                "summary": "Fetch workouts",
                "description": "Fetch all workouts for an authenticated user",
                "tags": [
                    "Workouts"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workouts fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "workouts": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Workout"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/workouts/library-exercises": {
            "get": {
                "summary": "Fetch filtered exercises",
                "tags": [
                    "Exercises"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Search term to filter exercises by name, bodyPart, equipment, or target"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "bodyPart": {
                                                "type": "string",
                                                "description": "Body part the exercise targets"
                                            },
                                            "equipment": {
                                                "type": "string",
                                                "description": "Equipment needed for the exercise"
                                            },
                                            "gifUrl": {
                                                "type": "string",
                                                "description": "URL of the exercise animation"
                                            },
                                            "id": {
                                                "type": "string",
                                                "description": "Exercise identifier"
                                            },
                                            "name": {
                                                "type": "string",
                                                "description": "Exercise name"
                                            },
                                            "target": {
                                                "type": "string",
                                                "description": "Target muscle group"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred while fetching the exercises"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/workouts/recommendations/{userId}": {
            "get": {
              "tags": [
                "Recommendations"
              ],
              "summary": "Get exercise recommendations for a user based on their fitness goals and workout history.",
              "description": "Returns a list of recommended exercises for a user.",
              "parameters": [
                {
                  "name": "userId",
                  "in": "path",
                  "description": "ID of the user to retrieve recommendations for.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                  "description": "An array of recommended exercises.",
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "description": "A message describing the purpose of the recommended exercises."
                        },
                        "exercises": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "description": "The name of the exercise."
                              },
                              "type": {
                                "type": "string",
                                "description": "The type of the exercise (e.g., strength, cardio, endurance)."
                              },
                              "muscleGroups": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                },
                                "description": "The muscle groups targeted by the exercise."
                              }
                            }
                          },
                          "description": "An array of exercise objects."
                        }
                      }
                    }
                  }
                },
                "404": {
                  "description": "User not found."
                }
              },
              "security": [
                {
                    "BearerAuth": []
                }
              ]
            }
          }          
    }
}